Update: createResponse.js / serialize.js
리스폰스 생성 및 직렬화 기능 업데이트
